% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkafka.R
\name{rkafka.closeProducer}
\alias{rkafka.closeProducer}
\title{function to shut down the producer}
\usage{
rkafka.closeProducer(producer)
}
\arguments{
\item{producer:producer(Java}{object)
!!Mandatory: Producer which is to be terminated}
}
\description{
function to shut down the producer
}
