% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkafka.R
\name{rkafka.createConsumer}
\alias{rkafka.createConsumer}
\title{Function to create consumer}
\usage{
rkafka.createConsumer(
  zookeeperConnect,
  topicName,
  groupId = "test-consumer-group",
  zookeeperConnectionTimeoutMs = "100000",
  consumerTimeoutMs = "10000",
  autoCommitEnable = "NULL",
  autoCommitInterval = "NULL",
  autoOffsetReset = "NULL"
)
}
\arguments{
\item{zookeeperConnect}{!!Mandatory:Zookeeper connection string comma separated
 host:port pairs, each corresponding to a zk server. e.g.
 "127.0.0.1:3000,127.0.0.1:3001,127.0.0.1:3002"
default:"127.0.0.1:2181"}

\item{groupId}{!!Mandatory:consumer group id default:test-consumer-group}

\item{zookeeperConnectionTimeoutMs}{!!Mandatory:timeout in ms for connecting to zookeeper
default:100000}

\item{consumerTimeoutMs}{!!Mandatory:Throw a timeout exception to the consumer if no
message is available for consumption after the specified
interval default:10000}

\item{autoCommitEnable}{--Optional:default:true If true, periodically commit to
ZooKeeper the offset of messages already fetched by the
consumer. This committed offset will be used when the process
fails as the position from which the new consumer will begin.}

\item{autoOffsetReset}{--Optional:default:largest
 smallest : automatically reset the offset to the smallest offset
largest : automatically: reset the offset to the largest offset
anything else: throw exception to the consumer}

\item{topicName:}{!!Mandatory: Name of the topic from where to read messages}

\item{autoCommitIntervalMs}{--Optional:default:601000 The frequency in ms that the
consumer offsets are committed to zookeeper.}
}
\description{
Function to create consumer
}
