% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkafka.R
\name{rkafka.createSimpleConsumer}
\alias{rkafka.createSimpleConsumer}
\title{Function to start Simple Consumer}
\usage{
rkafka.createSimpleConsumer(
  kafkaServerURL,
  kafkaServerPort,
  connectionTimeOut,
  kafkaProducerBufferSize,
  clientId
)
}
\arguments{
\item{kafkaServerURL}{: URL of the KAFKA server}

\item{kafkaServerPort}{:Port number of the KAFKA server}

\item{connectionTimeOut}{: Connection Timeout in ms}

\item{clientId}{: ID Of the client}

\item{kafkaProducerBufferSize:}{Buffer size}
}
\description{
Creates a KAFKA simple consumer
}
