% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkafka.R
\name{rkafka.read}
\alias{rkafka.read}
\title{Reads messages from the topic passed as parameter while creating the consumer object one at a time.Waits for a time
specified by consumer timeout property to check for new messages, else returns ""
	  @param ConsumerObj: Consumer through which messages are to be read(Java Object)
	  @return String: next available message}
\usage{
rkafka.read(ConsumerObj, verbose = FALSE)
}
\description{
Reads messages from the topic passed as parameter while creating the consumer object one at a time.Waits for a time
specified by consumer timeout property to check for new messages, else returns ""
	  @param ConsumerObj: Consumer through which messages are to be read(Java Object)
	  @return String: next available message
}
