% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkafka.R
\name{rkafka.readFromSimpleConsumer}
\alias{rkafka.readFromSimpleConsumer}
\title{Function to return messages read by Simple Consumer one at a time}
\usage{
rkafka.readFromSimpleConsumer(SimpleConsumerObj)
}
\description{
Function to return messages read by Simple Consumer one at a time
}
