% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkafka.R
\name{rkafka.readPoll}
\alias{rkafka.readPoll}
\title{Reads messages from the topic passed as parameter while creating the consumer object in batch.Waits for a time
specified by consumer timeout property and returns all the messages it read.}
\usage{
rkafka.readPoll(ConsumerObj, verbose = FALSE)
}
\arguments{
\item{ConsumerObj:}{Consumer through which messages are to be read(Java Object)}
}
\value{
String[]: Array of Strings
}
\description{
Reads messages from the topic passed as parameter while creating the consumer object in batch.Waits for a time
specified by consumer timeout property and returns all the messages it read.
}
