% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkafka.R
\name{rkafka.receiveFromSimpleConsumer}
\alias{rkafka.receiveFromSimpleConsumer}
\title{Function for consumer to read messages from topic}
\usage{
rkafka.receiveFromSimpleConsumer(
  SimpleConsumerObj,
  topicName,
  partition,
  Offset,
  msgReadSize
)
}
\arguments{
\item{Offset}{:Offset Number}

\item{msgReadSize}{: Size of the message to be read/}

\item{topicName:}{Name of the topic from where to read messages}

\item{partition:}{Partition Number}
}
\description{
Function for consumer to read messages from topic
}
