% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkafka.R
\name{rkafka.send}
\alias{rkafka.send}
\title{Function for producer to send message}
\usage{
rkafka.send(producer, topicName, ip, message, verbose = FALSE)
}
\arguments{
\item{producer:producer(Java}{object)
!!Mandatory: Producer through which messages are to be sent}

\item{topicName:String}{!!Mandatory: Topic to which messages are to be sent. If topicName doesn't exist, new topic is created}

\item{ip:String}{!!Mandatory: ip on which producer is running}

\item{message:String}{!!Mandatory: message to be sent}
}
\description{
Definition of the arguments
}
